代码清理：删除AutoFishingTrigger中被注释的私有变量；行为树扩展方法移动到单独的文件中
